+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "ihnatiev-site-v0.6.19c7-r1";
+const CACHE_NAME = "ihnatiev-site-v0.6.19c8-r1";
 const APP_SHELL = [
   "./",
   "./index.html",
@@ -16,6 +16,7 @@ const APP_SHELL = [
   "./js/youtube-feed.js",
   "./js/mobile-navigation.js",
   "./js/contact-page.js",
+  "./js/visitor-counter.js",
   "./app.js",
   "./manifest.webmanifest",
   "./files/media/icon-192.png",